Combine List.cs and IDictionary.cs; other feedback
