+++ src/libraries/System.Text.Json/src/System.Text.Json.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <TargetFrameworks>$(NetCoreAppCurrent);netstandard2.0;netcoreapp3.0;net461</TargetFrameworks>
@@ -58,7 +58,6 @@
     <Compile Include="System\Text\Json\Node\JsonObject.Dynamic.cs" />
     <Compile Include="System\Text\Json\Node\JsonObject.IDictionary.cs" />
     <Compile Include="System\Text\Json\Node\JsonObject.KeyCollection.cs" />
-    <Compile Include="System\Text\Json\Node\JsonObject.List.cs" />
     <Compile Include="System\Text\Json\Node\JsonObject.ValueCollection.cs" />
     <Compile Include="System\Text\Json\Node\JsonValue.cs" />
     <Compile Include="System\Text\Json\Node\JsonValueOfT.cs" />
